Fixed some minor problems
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@1066 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/SVN/SVNUrl.cpp
@@ -225,7 +225,7 @@ CString SVNUrl::GetTextFromRev(const SVNRev& revision)
 	}
 	else
 	{
-		rev_text.Format(_T("%u"), revision);
+		rev_text.Format(_T("%u"), (LONG)revision);
 	}
 
 	return rev_text;
